node_examples: renamed to examples and add a new example for custom tasks
