+++ docs/examples/tsconfig.eslint.json
