+++ docs/examples/tsconfig.json
