--- docs/node_example/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@epfml/disco-node-example",
-  "private": true,
-  "description": "In `main.ts` we give a brief example of discojs, in it we run two clients training jointly via federated learning. It trains  on a few examples of the [face task](https://www.kaggle.com/datasets/frabbisw/facial-age), the samples are already stored in the  repo and so it is not necessary to download any additional data.",
-  "main": "main.ts",
-  "scripts": {
-    "start": "npm run build && node dist/main.js",
-    "build": "tsc",
-    "lint": "npx eslint --max-warnings 0 .",
-    "test": ": nothing"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "@epfml/disco-server": "*",
-    "@epfml/discojs-node": "*"
-  },
-  "devDependencies": {
-    "ts-node": "10"
-  }
-}
